ARM: mstar: msc313 pm intc driver
This adds a driver for the PM interrupt *controller* found in
MStar/SigmaStar ARMv7 SoCs.

The PM parts of these chips is an always on domain that runs
even when the ARM CPU is powered down so that decoding an IR
command or similar can wake the ARM CPU up and boot the main
OS. It seems this part of the chip contains an 8051
micro-controller but I haven't been able to confirm that.

Controller is really a stretch here as it's basically a single
register that collects the interrupt status for other blocks in
the PM area, like the PM GPIO, that have their own interrupt
mask/clear/type bits.

The main function seems to be OR'ing together the interrupts
from the PM blocks to trigger a single interrupt that is wired
to the non-PM part of the chip so that hardware in the PM side
can be used when the ARM CPU is running.

In this driver when the upstream interrupt fires the status
register is used to trigger the right interrupt with PM.
The driver that is consuming the interrupt is responsible
for actually handling it.

Signed-off-by: Daniel Palmer <daniel@0x0f.com>