ARM: dts: mstar: Add the Wireless Tag IDO-SBC2D06-V1B-22W
The Wireless Tag IDO-SBC2D06-V1B-22W[1] is an SBC powered by SDD202D with a
dual Ethernet and a connector for a 4-inch and 7-inch display. It embeds
a System-On-Module IDO-SOM2D01[2] with an Mstar SSD202 SoC dual-core Cortex-A7
CPU @ 1.2Ghz , 2D GPU, 128 MB DDR3 (on-chip) and a 256MB SPI NAND flash.
This commit adds basic definition for this board.

1. http://linux-chenxing.org/infinity2/ido-sbc2d06
2. http://www.wireless-tag.com/portfolio/ido-som2d01

Signed-off-by: Romain Perier <romain.perier@gmail.com>