mfd: syscon: Add dt property to force using raw_spinlock
The MStar/SigmaStar ARMv7 SoCs have interrupt status bits
inside of lump'o'registers that is handled as a syscon
in devicetree.

The irqchip driver that handles these interrupts will
upset lockdep if normal spinlocks are used by regmap.

So add a property to specify that a syscon needs to
use raw spinlocks so that stuff like irqchip drivers
can be layered on top.

Signed-off-by: Daniel Palmer <daniel@0x0f.com>